add http chroma db driver

+++ memoflow/app/diary_log/controllers.py
@@ -58,11 +58,13 @@ class DiaryLog(wsgi.Application):
         que_string: List[str] = self.diary_log_api.get_que_string_from_content(
             processed_block_content)
         if que_string:
-            self.asyn_task_api.asyn_add_texts_to_vector_db(texts=que_string,
+            self.asyn_task_api.asyn_add_texts_to_vector_db_coll(texts=que_string,
                                                            metadatas=[{
                                                                "tags":
                                                                ','.join(tags)
                                                            }])
+            # self.vector_db_api.add_texts(texts=que_string,
+            #                              metadatas=[{"tags": ','.join(tags) }])
 
         # 保存到本地数据库
         self.diary_db_api.save_log(processed_content, tags)
@@ -193,8 +195,9 @@ class DiaryLog(wsgi.Application):
         if search_data:
             search_object_result = self.vector_db_api.search_texts(
                 query=search_data, top_k=10)
-        for item in search_object_result:
-            search_result.append(item.page_content)
+        for result in search_object_result:
+            for item in result['page_content']:
+                search_result.append(item)
         return json.dumps({"search_result": search_result})
 
     def update_all_que_to_vector_db(self, req):
@@ -212,10 +215,11 @@ class DiaryLog(wsgi.Application):
         # convert id into str
         ids = [str(log[0]) for log in slice_log_list]
 
-        all_ids = self.vector_db_api.get_collection_ids().get('ids', None)
+        all_ids = self.vector_db_api.get_vector_db_coll_all_ids().get('ids', None)
         LOG.info("length of all_ids: %s" % len(all_ids))
         # delete all ids
-        self.vector_db_api.delete_items_by_ids(ids=all_ids)
+        if all_ids:
+            self.vector_db_api.delete_items_by_ids(ids=all_ids)
 
         self.vector_db_api.add_texts(texts=que_list,
                                      metadatas=[{
@@ -245,8 +249,8 @@ class DiaryLog(wsgi.Application):
         }
         return json.dumps(result)
 
-    def get_collection_ids(self, req):
-        result_object = self.vector_db_api.get_collection_ids()
+    def get_vector_db_coll_all_ids(self, req):
+        result_object = self.vector_db_api.get_vector_db_coll_all_ids()
         collection_size = self.vector_db_api.get_collection_size()
         result = {
             "ids": result_object.get('ids', None),
@@ -264,7 +268,7 @@ class DiaryLog(wsgi.Application):
         return "success"
 
     def delete_all_collection_item(self, req):
-        all_collection_ids: List = self.vector_db_api.get_collection_ids().get(
+        all_collection_ids: List = self.vector_db_api.get_vector_db_coll_all_ids().get(
             'ids', None)
         if all_collection_ids:
             self.vector_db_api.delete_items_by_ids(ids=all_collection_ids)
